Merge in subtree for squashed merges
Instead of using git rebase with the subtree strategy, this performs a
merge on the subrepo/remote commits. The resulting tree is then grafted
into the main repo.

This matches with the behaviour of git-subrepo, gives more control over
what actually happens, and ensures conflict resolution will need to
happen only once.